Move finding_querys(), severity_count(), and endpoint_querys() out of Metric app's view.py and into queries.py. Also, because pylint was used to assist in this movement, this also clears some pylint warnings.
